+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useAkeneoTheme';
 export * from './useAutoFocus';
+export * from './useShortcut';