+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/hooks/useAutoFocus.ts
@@ -1,6 +1,6 @@
 import {RefObject, useCallback, useEffect} from 'react';
 
-const useAutoFocus = (ref: RefObject<HTMLInputElement>): (() => void) => {
+const useAutoFocus = (ref: RefObject<HTMLInputElement | HTMLTextAreaElement>): (() => void) => {
   const focus = useCallback(() => {
     if (ref.current !== null) ref.current.focus();
   }, [ref]);